// src/pages/search/interface.ts
export interface DogResponse {
  age?: string;
  animalType?: string;
  birthdate?: string;
  breed?: string;
  expirationDate?: string;
  findPet: boolean;
  id?: string;
  image?: string;
  loading: boolean;
  name?: string;
  owner?: string;
  registerDate?: string;
  registerState?: string;
  species?: string;
  weight?: string;
}